# src/server/swagger.json
{
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API for managing a priority queue",
    "title": "Priority Queue API",
    "version": "1.0.0"
  },
  "openapi": "3.0.4",
  "paths": {
    "/confirm": {
      "post": {
        "description": "Confirm a reservation by providing the reservation ID",
        "method": "post",
        "path": "/confirm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "description": "Reservation ID to confirm",
                "format": null,
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Reservation confirmed"
          },
          "400": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Forbidden"
          },
          "405": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Confirm a reservation"
      }
    },
    "/dequeue": {
      "get": {
        "description": "Dequeue an item from the priority queue",
        "method": "get",
        "parameters": [
          {
            "description": "Channel to dequeue from",
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "path": "/dequeue",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Dequeued item: {value}"
          },
          "204": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "No Content"
          },
          "400": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Forbidden"
          },
          "405": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Dequeue an item"
      }
    },
    "/enqueue": {
      "post": {
        "description": "Enqueue an item to the priority queue. The item is provided in the request body as a string (which can be a JSON object).",
        "method": "post",
        "parameters": [
          {
            "description": "Priority of the item",
            "in": "query",
            "name": "prio",
            "required": false,
            "schema": {
              "format": "float",
              "type": "number"
            }
          },
          {
            "description": "Channel to enqueue the item to",
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "description": "Timestamp in RFC3339 format specifying when the item becomes valid",
            "in": "query",
            "name": "notbefore",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "path": "/enqueue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "description": "Item to enqueue (string or JSON object)",
                "format": null,
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Item enqueued"
          },
          "400": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Forbidden"
          },
          "405": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Enqueue an item"
      }
    },
    "/reserve": {
      "get": {
        "description": "Dequeue an item from the priority queue with a reservation ID",
        "method": "get",
        "parameters": [
          {
            "description": "Channel to dequeue from",
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "path": "/reserve",
        "responses": {
          "200": {
            "content": {
              "map[string]string": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "{object}"
          },
          "204": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "No Content"
          },
          "400": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Forbidden"
          },
          "405": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Dequeue an item with reservation"
      }
    },
    "/reset": {
      "post": {
        "description": "Reset the priority queue",
        "method": "post",
        "path": "/reset",
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Queue reset successfully"
          },
          "403": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Forbidden"
          },
          "405": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Reset the queue"
      }
    },
    "/size": {
      "get": {
        "description": "Returns the number of items in the queue for a specified channel",
        "method": "get",
        "parameters": [
          {
            "description": "Channel to get the size of",
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "path": "/size",
        "responses": {
          "200": {
            "content": {
              "map[string]int": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "{object}"
          },
          "400": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Get the size of the queue"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}